Refactor: Move commonBundle to the LuviAppBundle module
